Added ShopListDialog, Added edit text and spinner in it, Changes to other dialog's buttons
