+++ app/src/main/java/com/fwrdgrp/recipesaving/ui/fragmentdialogs/AddShopListDialogFragment.kt
@@ -1,0 +1,48 @@
+package com.fwrdgrp.recipesaving.ui.fragmentdialogs
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
+import androidx.fragment.app.DialogFragment
+import com.fwrdgrp.recipesaving.data.models.shopping.Store
+import com.fwrdgrp.recipesaving.databinding.FragmentAddShopListDialogBinding
+
+class AddShopListDialogFragment : DialogFragment() {
+    private lateinit var binding: FragmentAddShopListDialogBinding
+    private lateinit var spinnerAdapter: ArrayAdapter<Store>
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentAddShopListDialogBinding.inflate(layoutInflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setupSpinner()
+    }
+
+    fun setupSpinner() {
+        binding.run {
+            spinnerAdapter = ArrayAdapter(
+                requireContext(),
+                android.R.layout.simple_spinner_item,
+                //Items here
+            )
+            spinnerAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+            spStoreName.adapter = spinnerAdapter
+            spStoreName.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                override fun onItemSelected(parent: AdapterView<*>?, view: View?, position: Int, id: Long) {
+                    val selected = spinnerAdapter.getItem(position) ?: return
+                    //Selected thing here
+                }
+                override fun onNothingSelected(parent: AdapterView<*>?) {}
+            }
+        }
+    }
+}
